Laser Rebalance (#310)
* BoS Pass #1
- - -
Provides Knights with the ability to wear PA, in line with the codex.
- - -

* Yuppers.

* Various Corrections
- - -
Permits proper BoS spawns.
Radio functionality for all map areas.
Minor rename to change a door in the BoS bunker. Most of the doors still need renames.

* Laser Rebalance
- - -
Significantly buffs lasers from the sorry state they're in.
AER14 is now the baseline for the AER9, with the AER12 getting a minor buff, alongside the AER14.

Additionally, beam weaponry, in regards to the AER9 onwards, alongside the AEP7, are now capable of wounding(bonus) just like the Wattz2k. They no longer inherit the negative wounding value of standard beams from base SS13.

Further, the armour penetration has gotten a flat increase across the board for beam weapons that you'll find in game, in most cases. Some weapons are untouched.

Lastly, fixes a minor typo with the minigun.
- - -