Improved Hazmat Suits Part 1
This is the first of 2 prs to make hazmat suits better, and actually worth using. It removes the slowdown of the suit, increases the melee resistance of the armor itself by a bit, acid proofs it, makes the helmet part of the suit instead of a separate object, and lets you put it in your backpack.