Minor minigun balance changes
I've consistently seen a majority of people refer to the minigun as a "Noob trap" and generally say that it's a poor gun. While I've had success with it in PvE, I've noticed in PvP, it tends to fall flat on its face unless you use it as a shotgun to surprise people in CQC. Anyone who knows what they're up against, and is prepared can generally abuse the slowdown to kill you with ease.

While it wouldn't make sense to completely take away all weaknesses from it, I figured I'd see about giving a very minor change to the two points people tend to mention most in regards to its balance.

I'm hoping these changes are light-handed, as, I don't intend to remove its weaknesses, but, simply lessen them. I debated a bit back and forth on if the heat diffusion should be touched at all, and, if I should change the slowdown less, but, in the, end, felt like that might be TOO light on things

Compared to the laser minigun, the ballistic minigun has double the slowdown, more spread, and, can't easily reload in the field due to the inability to carry around ammo boxes in a belt

The laser minigun might also benefit from a heat diffusion buff, but, aside from that is generally better than the minigun to the point of being a straight upgrade, so, I've left it untouched