Hee-Hoo Carl Loadout Abuse - - - Fluff:  - Added my own loadout. Funny abuse time woohoo. Armour has next to no protection outside of utility, and is used for a ghoul character as is, so that's another thing to tack on. It's thematically centred around being hazard equipment and is practically useless in PvP except for the fire immunity, which isn't really used outside of the NCR as is.
Helmet trades practically all protection for utility, and retains flash resist and the NVG overlay, which is worthless for ghouls anyway and not that great as is.

General armour layout, for the suit and helmet, with the rest having identical without actual protection values aside from utility (The 100 row.):
 - melee/bullet/laser = 5
 - energy = 65
 - bomb = 5
 - bio/rad/fire/acid = 100
 - wound = 15

The armour is able to be reinforced with plating, but that can easily be removed.
The equipment itself consists of:
 - Suit
 - Helmet
 - Hazard gloves / boots / uniform (None have a unique name, instead having a unique description.)
 - Old dogtags.
 - Compliance Regulator holster.
 - Nuke core container and screwdriver.
- - -