Bighorn Power Plant (Properly, this time)
Cleaned up the commits, used Map Merge 2 to actually merge the map properly. Adds a power plant to Bighorn, adds a construction area, sets Bighorn, the Followers, and Bighorn Underground to require power, adds access to Followers and Bighorners, puts the Schoolhouse on the Follower Grid, changes the Schoolhouse to match Follower construction.