water boil fix
was impossible to make purified water with carbon and iodine because salt had a lower required temp. slightly raised salt making temp and lowered boiled to purified water. this is important since WATER is used in chem recipies and other things. without a dispenser can such to get in the wastes