Bug fix
this was busted before I even touched it and im fucking glad I did. this is what has been breaking defibs. when you toxin removed someone it removes the toxin but doesnt count to the overall heal. its why you had to melee damage them after removing it for the game to register there new health state when dead. I gave a very shitty fix to it but it works. set it to give light light light .01 brute damage on each step. so the game will register the over all health change